+++ packages/webgl-components/src/loading/image-loader.ts
@@ -1,4 +1,3 @@
-import { LoadingStatus } from './asset';
 import Loader, { LoaderSettings, LoadingEnvironment } from './loader';
 import LoaderManager from './loader-manager';
 
@@ -17,11 +16,6 @@ export default class ImageLoader extends Loader {
 
     manager.add(this);
 
-    const onError = (error: string | Event) => {
-      this.asset.status = LoadingStatus.Error;
-      this.emit('error', error, `Failed to load ${this.asset.src}`);
-    };
-
     if (this.settings.environment === LoadingEnvironment.Worker && this.settings.preferWebWorker) {
       fetch(this.asset.src)
         .then((response) => {
@@ -31,15 +25,14 @@ export default class ImageLoader extends Loader {
               const type = blob.type.split('/')[1];
               if (/(gif|jpe?g|tiff?|png|webp)$/i.test(type)) {
                 this.asset.data = URL.createObjectURL(blob);
-                this.asset.status = LoadingStatus.Loaded;
                 this.emit('loaded', this.asset);
               } else {
-                onError(`Image type not supported: ${type}`);
+                this.onError(`Image type not supported: ${type}`);
               }
             })
-            .catch(onError);
+            .catch(this.onError);
         })
-        .catch(onError);
+        .catch(this.onError);
     } else {
       const image = new Image();
 
@@ -47,7 +40,7 @@ export default class ImageLoader extends Loader {
         this.asset.data = image;
         this.emit('loaded', this.asset);
       };
-      image.onerror = onError;
+      image.onerror = this.onError;
       image.src = this.asset.src;
     }
   };
